refactor CreatorLockForms test
