+++ unlock-app/src/__tests__/components/creator/CreatorLocks.test.js
@@ -1,0 +1,57 @@
+import React from 'react'
+import * as rtl from 'react-testing-library'
+import { Provider } from 'react-redux'
+
+import { CreatorLocks } from '../../../components/creator/CreatorLocks'
+import createUnlockStore from '../../../createUnlockStore'
+
+jest.mock('next/link', () => {
+  return ({ children }) => children
+})
+
+describe('CreatorLocks', () => {
+  it('should display form when create lock button is clicked', () => {
+    const store = createUnlockStore({
+      account: {},
+    })
+
+    const wrapper = rtl.render(
+      <Provider store={store}>
+        <CreatorLocks />
+      </Provider>
+    )
+
+    expect(wrapper.queryByValue('New Lock')).toBeNull()
+    expect(wrapper.queryByText('Submit')).toBeNull()
+
+    const createButton = wrapper.getByText('Create Lock')
+    rtl.fireEvent.click(createButton)
+
+    expect(wrapper.queryByValue('New Lock')).not.toBeNull()
+    expect(wrapper.queryByText('Submit')).not.toBeNull()
+  })
+
+  it('should disappear when cancel button is clicked', () => {
+    const store = createUnlockStore({
+      account: {},
+    })
+
+    let wrapper = rtl.render(
+      <Provider store={store}>
+        <CreatorLocks />
+      </Provider>
+    )
+
+    let createButton = wrapper.getByText('Create Lock')
+    rtl.fireEvent.click(createButton)
+
+    expect(wrapper.queryByValue('New Lock')).not.toBeNull()
+    expect(wrapper.queryByText('Submit')).not.toBeNull()
+
+    let cancelButton = wrapper.getByText('Cancel')
+    rtl.fireEvent.click(cancelButton)
+
+    expect(wrapper.queryByValue('New Lock')).toBeNull()
+    expect(wrapper.queryByText('Submit')).toBeNull()
+  })
+})
